+++ examples/simple/runserver-bootstrap3-theme-captcha.sh
@@ -1,2 +1,2 @@
 #workon fobi
-./manage.py runserver 0.0.0.0:8004 --settings=settings_bootstrap3_theme_captcha  --traceback -v 3
+./manage.py runserver 0.0.0.0:8000 --settings=settings_bootstrap3_theme_captcha  --traceback -v 3