+++ examples/simple/runserver-bootstrap3-theme-feincms-integration.sh
@@ -1,0 +1,2 @@
+#workon fobi
+./manage.py runserver 0.0.0.0:8000 --settings=settings_bootstrap3_theme_feincms --traceback -v 3