+++ examples/simple/runserver-feincms-integration.sh
@@ -1,2 +1,0 @@
-#workon fobi
-./manage.py runserver 0.0.0.0:8005 --settings=settings_feincms --traceback -v 3