+++ examples/simple/runserver-foundation5-theme-feincms-integration.sh
@@ -1,0 +1,3 @@
+#workon fobi
+#./manage.py collectstatic --noinput --settings=settings_foundation5_theme_feincms --traceback -v 3
+./manage.py runserver 0.0.0.0:8001 --settings=settings_foundation5_theme_feincms --traceback -v 3