+++ src/fobi/__init__.py
@@ -1,6 +1,6 @@
 __title__ = 'django-fobi'
-__version__ = '0.1.5'
-__build__ = 0x000006
+__version__ = '0.1.6'
+__build__ = 0x000007
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'