+++ src/fobi/contrib/apps/feincms_integration/defaults.py
@@ -2,6 +2,13 @@ __title__ = 'fobi.contrib.apps.feincms_integration.defaults'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
-__all__ = ('WIDGET_FORM_SENT_GET_PARAM',)
+__all__ = (
+    'WIDGET_FORM_SENT_GET_PARAM', 'FORM_TEMPLATE_CHOICES',
+    'SUCCESS_PAGE_TEMPLATE_CHOICES',
+)
 
 WIDGET_FORM_SENT_GET_PARAM = 'sent'
+
+FORM_TEMPLATE_CHOICES = []
+
+SUCCESS_PAGE_TEMPLATE_CHOICES = []