+++ src/fobi/contrib/apps/feincms_integration/widgets.py
@@ -10,7 +10,7 @@ from django.template.loader import render_to_string
 from django.template import RequestContext
 from django.contrib import messages
 from django.shortcuts import redirect
-from django.core.urlresolvers import reverse
+#from django.core.urlresolvers import reverse
 from django.conf import settings
 
 from fobi.dynamic import assemble_form_class
@@ -26,12 +26,60 @@ from fobi.constants import (
 from fobi.contrib.apps.feincms_integration.settings import (
     WIDGET_FORM_SENT_GET_PARAM
     )
+from fobi.contrib.apps.feincms_integration.helpers import (
+    get_form_template_choices, get_success_page_template_choices
+)
 
 class FobiFormWidget(models.Model):
     """
     Widget for to FeinCMS.
+
+    :property fobi.models.FormEntry form_entry: Form entry to be rendered.
+    :property str template: If given used for rendering the form.
     """
     form_entry = models.ForeignKey('fobi.FormEntry', verbose_name=_("Form"))
+
+    form_template_name = models.CharField(
+        _("Form template name"), max_length=255, null=True, blank=True,
+        choices=get_form_template_choices(),
+        help_text=_("Template to render the form with.")
+        )
+
+    hide_form_title = models.BooleanField(
+        _("Hide form title"), default=False,
+        help_text=_("If checked, no form title is shown.")
+        )
+
+    form_title = models.CharField(
+        _("Form title"), max_length=255, null=True, blank=True,
+        help_text=_("Overrides the default form title.")
+        )
+
+    form_submit_button_text = models.CharField(
+        _("Submit button text"), max_length=255, null=True, blank=True,
+        help_text=_("Overrides the default form submit button text.")
+        )
+
+    success_page_template_name = models.CharField(
+        _("Success page template name"), max_length=255, null=True, blank=True,
+        choices=get_success_page_template_choices(),
+        help_text=_("Template to render the success page with.")
+        )
+
+    hide_success_page_title = models.BooleanField(
+        _("Hide success page title"), default=False,
+        help_text=_("If checked, no success page title is shown.")
+        )
+
+    success_page_title = models.CharField(
+        _("Succes page title"), max_length=255, null=True, blank=True,
+        help_text=_("Overrides the default success page title.")
+        )
+
+    success_page_text = models.TextField(
+        _("Succes page text"), null=True, blank=True,
+        help_text=_("Overrides the default success page text.")
+        )
 
     class Meta:
         abstract = True
@@ -60,7 +108,7 @@ class FobiFormWidget(models.Model):
         :param django.http.HttpRequest request:
         :return django.http.HttpResponse | str:
         """
-        template_name = None # TODO: fix
+        template_name = self.form_template_name or None
 
         # Handle public/non-public forms. If form requires user authentication
         # redirect to login form with next parameter set to current request
@@ -131,7 +179,7 @@ class FobiFormWidget(models.Model):
                     request,
                     _('Form {0} was submitted successfully.').format(self.form_entry.name)
                     )
-                #return redirect(reverse('fobi.form_entry_submitted', args=[self.form_entry.slug]))
+
                 return redirect(
                     "{0}?{1}={2}".format(request.path, WIDGET_FORM_SENT_GET_PARAM, \
                                          self.form_entry.slug)
@@ -156,10 +204,13 @@ class FobiFormWidget(models.Model):
             'form': form,
             'form_entry': self.form_entry,
             'fobi_theme': theme,
+            'fobi_form_title': self.form_title,
+            'fobi_hide_form_title': self.hide_form_title,
+            'fobi_form_submit_button_text': self.form_submit_button_text
         }
 
         if not template_name:
-            template_name = theme.view_form_entry_ajax_template
+            template_name = theme.view_embed_form_entry_ajax_template
 
         self.rendered_output = render_to_string(
             template_name, context, context_instance=RequestContext(request)
@@ -172,17 +223,20 @@ class FobiFormWidget(models.Model):
         :param django.http.HttpRequest request:
         :return str:
         """
-        template_name = None # TODO
+        template_name = self.success_page_template_name or None
 
         theme = get_theme(request=request, as_instance=True)
 
         context = {
             'form_entry': self.form_entry,
             'fobi_theme': theme,
+            'fobi_hide_success_page_title': self.hide_success_page_title,
+            'fobi_success_page_title': self.success_page_title,
+            'fobi_success_page_text': self.success_page_text,
         }
 
         if not template_name:
-            template_name = theme.form_entry_submitted_template
+            template_name = theme.embed_form_entry_submitted_ajax_template
 
         self.rendered_output = render_to_string(
             template_name, context, context_instance=RequestContext(request)