+++ CHANGELOG.rst
@@ -1,5 +1,23 @@
 Release history
 =====================================
+0.1.6
+-------------------------------------
+2014-10-25
+
+- Minor improvements in the theming API. From now on the
+  ``view_embed_form_entry_ajax_template`` template would be used
+  when integrating the form rendering from other products (for example,
+  a CMS page, which has a widget which references the form object. If
+  that property is left empty, the ``view_form_entry_ajax_template``
+  is used. For a success page the ``embed_form_entry_submitted_ajax_template``
+  template would be used.
+- Functional improvements of the FeinCMS integration (the widget). If you
+  have used the FeinCMS widget of earlier versions, you likely want to update 
+  to this one. From now on you can select a custom form title and the button
+  text, as well as provide custom success page title and the success  message;
+  additionally, it has been made possible to hide the form- or success-page-
+  titles.
+
 0.1.5
 -------------------------------------
 2014-10-23